finished

+++ srcs/requirements/wordpress/conf/wp_conf.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+
 #---------------------------------------------------wp installation---------------------------------------------------#
 # wp-cli installation
 curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
@@ -7,24 +8,42 @@ chmod +x wp-cli.phar
 # wp-cli move to bin
 mv wp-cli.phar /usr/local/bin/wp
 
+# create wordpress dir
+mkdir -p /var/www/wordpress
 # go to wordpress directory
 cd /var/www/wordpress
 # give permission to wordpress directory
-chmod -R 755 /var/www/wordpress/
+chmod -R 755 /var/www/wordpress
 # change owner of wordpress directory to www-data
 chown -R www-data:www-data /var/www/wordpress
-#---------------------------------------------------wp installation---------------------------------------------------##---------------------------------------------------wp installation---------------------------------------------------#
 
 # download wordpress core files
 wp core download --allow-root
+
+#---------------------------------------------------wp-config creation---------------------------------------------------#
+
 # create wp-config.php file with database details
 wp core config --dbhost=mariadb:3306 --dbname="$MYSQL_DB" --dbuser="$MYSQL_USER" --dbpass="$MYSQL_PASSWORD" --allow-root
+
+# Check if wp-config.php was created successfully
+if [ ! -f /var/www/wordpress/wp-config.php ]; then
+  echo "wp-config.php not created!"
+  exit 1
+fi
+
+# Modify wp-config.php file permissions
+chown www-data:www-data /var/www/wordpress/wp-config.php
+chmod 644 /var/www/wordpress/wp-config.php
+
+#---------------------------------------------------wp installation---------------------------------------------------#
+
 # install wordpress with the given title, admin username, password and email
 wp core install --url="$DOMAIN_NAME" --title="$WP_TITLE" --admin_user="$WP_ADMIN_N" --admin_password="$WP_ADMIN_P" --admin_email="$WP_ADMIN_E" --allow-root
+
 #create a new user with the given username, email, password and role
 wp user create "$WP_U_NAME" "$WP_U_EMAIL" --user_pass="$WP_U_PASS" --role="$WP_U_ROLE" --allow-root
 
-#---------------------------------------------------php config---------------------------------------------------#
+#---------------------------------------------------php config--------------------------------------------------------#
 
 # change listen port from unix socket to 9000
 sed -i '36 s@/run/php/php7.4-fpm.sock@9000@' /etc/php/7.4/fpm/pool.d/www.conf
